fix: deferred gi, TODO: 延迟管线下要给 other camera 做一个基于 MRT 的延迟管线，以适配 Unity Editor Preview Window， reflect bake 等

+++ Shaders/ShaderLibrarys/UnityInput.hlsl
@@ -101,8 +101,8 @@ CBUFFER_START(UnityPerDraw)
 
     float4 unity_ProbesOcclusion;
 
-    half4 unity_SpecCube0_HDR;
-    half4 unity_SpecCube1_HDR;
+    float4 unity_SpecCube0_HDR;
+    float4 unity_SpecCube1_HDR;
 
     float4 unity_SpecCube0_BoxMax;
     float4 unity_SpecCube0_BoxMin;
